product done

+++ backend/product-service/src/main/java/com/novamart/product_service/service/ReviewService.java
@@ -1,5 +1,6 @@
 package com.novamart.product_service.service;
 
+import com.novamart.product_service.client.UserClient;
 import com.novamart.product_service.dto.ReviewRequest;
 import com.novamart.product_service.dto.UpdateReviewRequest;
 import com.novamart.product_service.model.Reviews;
@@ -16,8 +17,12 @@ import java.util.UUID;
 @Slf4j
 public class ReviewService {
     private final ReviewRepository reviewRepository;
+    private final UserClient userClient;
 
     public void addReview(ReviewRequest reviewRequest) {
+        if (!userClient.authenticateUser(reviewRequest.userId(), "accountType", "CUSTOMER")) {
+            throw new RuntimeException("User not authenticated");
+        }
         Reviews review = Reviews.builder()
                         .reviewId(UUID.randomUUID().toString())
                         .userId(reviewRequest.userId())
@@ -29,7 +34,6 @@ public class ReviewService {
                         .rating(reviewRequest.rating())
                         .build();
         reviewRepository.save(review);
-        log.info("add review");
     }
 
     public List<Reviews> getReviewByProductId(String productId) {
@@ -46,22 +50,29 @@ public class ReviewService {
 
     public Reviews updateReview(UpdateReviewRequest updateReviewRequest) {
         Reviews review = reviewRepository.findByReviewId(updateReviewRequest.reviewId());
+        if (!userClient.authenticateUser(review.getUserId(), "accountType", "CUSTOMER")) {
+            throw new RuntimeException("User not authenticated");
+        }
         review.setTitle(updateReviewRequest.title());
         review.setComment(updateReviewRequest.comment());
         review.setRating(updateReviewRequest.rating());
         review.setImageUrl(updateReviewRequest.imageUrl());
         reviewRepository.save(review);
-        log.info("update review");
         return review;
     }
 
     public void deleteReview(String reviewId) {
         Reviews review = reviewRepository.findByReviewId(reviewId);
+        if (!userClient.authenticateUser(review.getUserId(), "accountType", "CUSTOMER")) {
+            throw new RuntimeException("User not authenticated");
+        }
         reviewRepository.delete(review);
-        log.info("delete review");
     }
 
-    public void deleteAllReviews() {
+    public void deleteAllReviews(String userId) {
+        if (!userClient.authenticateUser(userId, "accountType", "ADMIN")) {
+            throw new RuntimeException("User not authenticated");
+        }
         reviewRepository.deleteAll();
     }
 }
